requirements.txt
fastapi
uvicorn[standard]
pyjwt
aiogram
requests
python-multipart
